+++ src/expand/expand_parameter/state/on_error_expand.c
@@ -1,27 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   on_success.c                                       :+:      :+:    :+:   */
+/*   on_error.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: surayama <surayama@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/02/01 22:38:01 by surayama          #+#    #+#             */
-/*   Updated: 2026/02/07 00:15:14 by surayama         ###   ########.fr       */
+/*   Created: 2026/02/01 22:42:03 by surayama          #+#    #+#             */
+/*   Updated: 2026/02/07 00:15:10 by surayama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../expand_variable_internal.h"
+#include "../expand_parameter_internal.h"
 #include "expand.h"
 
-t_list	*on_success_expand(t_expand_store *store, char *token)
+t_list	*on_error_expand(t_expand_store *store, char *token)
 {
-	t_list	*expanded_tokens;
-
 	(void)token;
-	if (!store)
-		return (NULL);
-	expanded_tokens = store->tokens;
-	store->tokens = NULL;
-	ft_lstclear(&(store->buffer), free);
-	return (expanded_tokens);
+	if (store)
+		free_store_expand(store);
+	return (NULL);
 }
