+++ src/expand/expand_parameter/store/free_store_expand.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../expand_variable_internal.h"
+#include "../expand_parameter_internal.h"
 #include "expand.h"
 
 void	free_store_expand(t_expand_store *store)
