+++ src/expand/expand_parameter/store/push_token_expand.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../expand_variable_internal.h"
+#include "../expand_parameter_internal.h"
 #include "expand.h"
 
 static char	*build_token_from_buffer(t_list *buffer)
